UPDATE refactor Dot component to seperate component
