+++ client/src/components/Map/Dot/Dot.css
@@ -1,0 +1,3 @@
+.dot_wrapper{
+
+}
