+++ client/src/components/Map/Dot/Dot.react.js
@@ -1,0 +1,26 @@
+import React from "react";
+
+import "./Dot.css";
+
+const Dot = props => {
+  const colors = {
+    housing: "#E83338",
+    education: "#ff9068",
+    agriculture: "#FFB75E",
+    health: "#8DC26F",
+    water: "#64b3f4",
+    nutrition: "#6441A5",
+    infancy: "#fc67fa"
+  };
+  return (
+    <div className="dot" style={{ background: colors[props.project.type] }}>
+      <div class="tooltip">
+        <div class="tooltiptext">{props.project.title}</div>
+        <img src="https://source.unsplash.com/random" />
+        <p className="tooltip-money">{props.project.intensity}$</p>
+      </div>
+    </div>
+  );
+};
+
+export default Dot;
