+++ client/src/components/Map/Filter/Filter.css
@@ -1,0 +1,3 @@
+.filter_wrapper{
+
+}
