+++ client/src/components/Map/Filter/Filter.react.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+import './Filter.css';
+
+class Filter extends React.Component {
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return (
+            <div className='filter_wrapper'>
+
+            </div>
+        );
+    }
+}
+
+export default Filter;
